React fragment
A fragment wraps other elements into a single top-level element without adding to the rendered output. Both Search elements should be visible in your browser now, with input field and label. So if you prefer to omit the wrapping <div> or <span> elements, substitute them with an empty tag that is allowed in JSX, and doesn’t introduce intermediate elements in our rendered HTML.